Suppression des fichier mal nommé doublon et actual merge
