+++ src/Media/Controllers/MediasController.ts
@@ -39,6 +39,10 @@ class MediasController extends AbstractController implements ControllerContract 
     public basepath(requestData:any) {
         return SuccessResponse.create({"basepath":"truetrue"}, StatusCodes.OK, "asdasdasd");
     }
+    
+    public uploadSingle(requestData:any) {
+        return SuccessResponse.create({"uploadSingle":"truetrue"}, StatusCodes.OK, "Uploading single media comming into dev!");
+    }
 }
 
 export default MediasController;
